services/hummingbot-gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
paho-mqtt==1.6.1
docker==7.1.0
urllib3<2.0.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
